#485 starting refactor to generalize all staging and agg sqlstrs
