+++ src/f10_etl/test_tran/test_idea_staging_tables2fiscal_staging_tables.py
@@ -22,6 +22,12 @@ from src.f04_gift.atom_config import (
     respect_bit_str,
 )
 from src.f07_fiscal.fiscal_config import (
+    fiscal_cashbook_str,
+    fiscal_deal_episode_str,
+    fiscal_timeline_hour_str,
+    fiscal_timeline_month_str,
+    fiscal_timeline_weekday_str,
+    fiscalunit_str,
     present_time_str,
     amount_str,
     hour_title_str,
@@ -31,25 +37,25 @@ from src.f07_fiscal.fiscal_config import (
     weekday_order_str,
     weekday_title_str,
 )
-from src.f08_pidgin.pidgin_config import event_int_str
-from src.f09_idea.idea_config import idea_number_str, get_idea_sqlite_types
+from src.f08_pidgin.pidgin_config import event_int_str, pidginunit_str
+from src.f09_idea.idea_config import (
+    idea_number_str,
+    get_idea_sqlite_types,
+    get_idea_config_dict,
+    idea_type_str,
+)
 from src.f09_idea.pandas_tool import get_pragma_table_fetchall, get_custom_sorted_list
 from src.f10_etl.fiscal_etl_tool import (
     FiscalPrimeObjsRef,
     FiscalPrimeColumnsRef,
 )
+from src.f10_etl.tran_sqlstrs import get_inconsistency_sqlstrs
 from src.f10_etl.transformers import (
     etl_aft_face_csv_files_to_fiscal_db,
     create_fiscal_tables,
     idea_staging_tables2fiscal_staging_tables,
     etl_fiscal_staging_tables_to_fiscal_csvs,
     set_fiscal_staging_error_message,
-    FISCALUNIT_INCONSISTENCY_SQLSTR,
-    FISCALDEAL_INCONSISTENCY_SQLSTR,
-    FISCALCASH_INCONSISTENCY_SQLSTR,
-    FISCALHOUR_INCONSISTENCY_SQLSTR,
-    FISCALMONT_INCONSISTENCY_SQLSTR,
-    FISCALWEEK_INCONSISTENCY_SQLSTR,
 )
 from src.f10_etl.examples.etl_env import get_test_etl_dir, env_dir_setup_cleanup
 from sqlite3 import connect as sqlite3_connect, Connection as sqlite3_Connection
@@ -61,84 +67,6 @@ def create_idea_staging_table(
     conn: sqlite3_Connection, tablename: str, columns_list: list[str]
 ):
     create_table_from_columns(conn, tablename, columns_list, get_idea_sqlite_types())
-
-
-def test_create_fiscal_tables_CreatesFiscalStagingTables(
-    env_dir_setup_cleanup,
-):
-    # ESTABLISH
-    with sqlite3_connect(":memory:") as fiscal_db_conn:
-        fis_objs = FiscalPrimeObjsRef()
-        fis_cols = FiscalPrimeColumnsRef()
-        assert db_table_exists(fiscal_db_conn, fis_objs.unit_agg_tablename) is False
-        assert db_table_exists(fiscal_db_conn, fis_objs.deal_agg_tablename) is False
-        assert db_table_exists(fiscal_db_conn, fis_objs.cash_agg_tablename) is False
-        assert db_table_exists(fiscal_db_conn, fis_objs.hour_agg_tablename) is False
-        assert db_table_exists(fiscal_db_conn, fis_objs.mont_agg_tablename) is False
-        assert db_table_exists(fiscal_db_conn, fis_objs.week_agg_tablename) is False
-        assert db_table_exists(fiscal_db_conn, fis_objs.unit_stage_tablename) is False
-        assert db_table_exists(fiscal_db_conn, fis_objs.deal_stage_tablename) is False
-        assert db_table_exists(fiscal_db_conn, fis_objs.cash_stage_tablename) is False
-        assert db_table_exists(fiscal_db_conn, fis_objs.hour_stage_tablename) is False
-        assert db_table_exists(fiscal_db_conn, fis_objs.mont_stage_tablename) is False
-        assert db_table_exists(fiscal_db_conn, fis_objs.week_stage_tablename) is False
-
-        # WHEN
-        create_fiscal_tables(fiscal_db_conn)
-
-        # THEN
-        assert db_table_exists(fiscal_db_conn, fis_objs.unit_agg_tablename)
-        assert db_table_exists(fiscal_db_conn, fis_objs.deal_agg_tablename)
-        assert db_table_exists(fiscal_db_conn, fis_objs.cash_agg_tablename)
-        assert db_table_exists(fiscal_db_conn, fis_objs.hour_agg_tablename)
-        assert db_table_exists(fiscal_db_conn, fis_objs.mont_agg_tablename)
-        assert db_table_exists(fiscal_db_conn, fis_objs.week_agg_tablename)
-
-        assert db_table_exists(fiscal_db_conn, fis_objs.unit_stage_tablename)
-        assert db_table_exists(fiscal_db_conn, fis_objs.deal_stage_tablename)
-        assert db_table_exists(fiscal_db_conn, fis_objs.cash_stage_tablename)
-        assert db_table_exists(fiscal_db_conn, fis_objs.hour_stage_tablename)
-        assert db_table_exists(fiscal_db_conn, fis_objs.mont_stage_tablename)
-        assert db_table_exists(fiscal_db_conn, fis_objs.week_stage_tablename)
-
-        fis_unit_agg_pragma = get_pragma_table_fetchall(fis_cols.unit_agg_columns)
-        fis_deal_agg_pragma = get_pragma_table_fetchall(fis_cols.deal_agg_columns)
-        fis_cash_agg_pragma = get_pragma_table_fetchall(fis_cols.cash_agg_columns)
-        fis_hour_agg_pragma = get_pragma_table_fetchall(fis_cols.hour_agg_columns)
-        fis_mont_agg_pragma = get_pragma_table_fetchall(fis_cols.mont_agg_columns)
-        fis_week_agg_pragma = get_pragma_table_fetchall(fis_cols.week_agg_columns)
-        fis_unit_stage_pragma = get_pragma_table_fetchall(fis_cols.unit_staging_columns)
-        fis_deal_stage_pragma = get_pragma_table_fetchall(fis_cols.deal_staging_columns)
-        fis_cash_stage_pragma = get_pragma_table_fetchall(fis_cols.cash_staging_columns)
-        fis_hour_stage_pragma = get_pragma_table_fetchall(fis_cols.hour_staging_columns)
-        fis_mont_stage_pragma = get_pragma_table_fetchall(fis_cols.mont_staging_columns)
-        fis_week_stage_pragma = get_pragma_table_fetchall(fis_cols.week_staging_columns)
-        cursor = fiscal_db_conn.cursor()
-        cursor.execute(f"PRAGMA table_info({fis_objs.unit_agg_tablename})")
-        assert fis_unit_agg_pragma == cursor.fetchall()
-        cursor.execute(f"PRAGMA table_info({fis_objs.deal_agg_tablename})")
-        assert fis_deal_agg_pragma == cursor.fetchall()
-        cursor.execute(f"PRAGMA table_info({fis_objs.cash_agg_tablename})")
-        assert fis_cash_agg_pragma == cursor.fetchall()
-        cursor.execute(f"PRAGMA table_info({fis_objs.hour_agg_tablename})")
-        assert fis_hour_agg_pragma == cursor.fetchall()
-        cursor.execute(f"PRAGMA table_info({fis_objs.mont_agg_tablename})")
-        assert fis_mont_agg_pragma == cursor.fetchall()
-        cursor.execute(f"PRAGMA table_info({fis_objs.week_agg_tablename})")
-        assert fis_week_agg_pragma == cursor.fetchall()
-
-        cursor.execute(f"PRAGMA table_info({fis_objs.unit_stage_tablename})")
-        assert fis_unit_stage_pragma == cursor.fetchall()
-        cursor.execute(f"PRAGMA table_info({fis_objs.deal_stage_tablename})")
-        assert fis_deal_stage_pragma == cursor.fetchall()
-        cursor.execute(f"PRAGMA table_info({fis_objs.cash_stage_tablename})")
-        assert fis_cash_stage_pragma == cursor.fetchall()
-        cursor.execute(f"PRAGMA table_info({fis_objs.hour_stage_tablename})")
-        assert fis_hour_stage_pragma == cursor.fetchall()
-        cursor.execute(f"PRAGMA table_info({fis_objs.mont_stage_tablename})")
-        assert fis_mont_stage_pragma == cursor.fetchall()
-        cursor.execute(f"PRAGMA table_info({fis_objs.week_stage_tablename})")
-        assert fis_week_stage_pragma == cursor.fetchall()
 
 
 def test_idea_staging_tables2fiscal_staging_tables_Scenario0_From_br00011_IdeaFile(
@@ -981,77 +909,31 @@ VALUES
 
 
 def test_GlobalVariablesForFiscal_inconsistency_queryReturns_sqlstrs():
+    # sourcery skip: extract-method, no-loop-in-tests
     # ESTABLISH
-    x_objs = FiscalPrimeObjsRef()
+    idea_config = get_idea_config_dict()
+    idea_config = {
+        x_category: category_config
+        for x_category, category_config in idea_config.items()
+        # if category_config.get(idea_type_str()) != pidginunit_str()
+        if category_config.get(idea_type_str()) == fiscalunit_str()
+    }
+
     exclude_cols = {"idea_number", "face_name", "event_int", "error_message"}
-    with sqlite3_connect(":memory:") as fiscal_db_conn:
-        create_fiscal_tables(fiscal_db_conn)
+    with sqlite3_connect(":memory:") as conn:
+        create_fiscal_tables(conn)
 
-        # WHEN
-        unit_tablename = x_objs.unit_stage_tablename
-        unit_focus_columns = {fiscal_title_str()}
-        generated_fiscalunit_sqlstr = create_inconsistency_query(
-            fiscal_db_conn, unit_tablename, unit_focus_columns, exclude_cols
-        )
-
-        # THEN
-        print(f"{generated_fiscalunit_sqlstr=}")
-        assert FISCALUNIT_INCONSISTENCY_SQLSTR == generated_fiscalunit_sqlstr
-
-        # WHEN
-        deal_tablename = x_objs.deal_stage_tablename
-        deal_focus_columns = {fiscal_title_str(), owner_name_str(), time_int_str()}
-        generated_fiscaldeal_sqlstr = create_inconsistency_query(
-            fiscal_db_conn, deal_tablename, deal_focus_columns, exclude_cols
-        )
-        # THEN
-        print(f"{generated_fiscaldeal_sqlstr=}")
-        assert FISCALDEAL_INCONSISTENCY_SQLSTR == generated_fiscaldeal_sqlstr
-
-        # WHEN
-        cash_tablename = x_objs.cash_stage_tablename
-        cash_focus_columns = {
-            fiscal_title_str(),
-            owner_name_str(),
-            acct_name_str(),
-            time_int_str(),
-        }
-        generated_fiscalcash_sqlstr = create_inconsistency_query(
-            fiscal_db_conn, cash_tablename, cash_focus_columns, exclude_cols
-        )
-        # THEN
-        print(f"{generated_fiscalcash_sqlstr=}")
-        assert FISCALCASH_INCONSISTENCY_SQLSTR == generated_fiscalcash_sqlstr
-
-        # WHEN
-        hour_tablename = x_objs.hour_stage_tablename
-        focus_columns = {fiscal_title_str(), hour_title_str()}
-        generated_fiscalhour_sqlstr = create_inconsistency_query(
-            fiscal_db_conn, hour_tablename, focus_columns, exclude_cols
-        )
-        # THEN
-        print(f"{generated_fiscalhour_sqlstr=}")
-        assert FISCALHOUR_INCONSISTENCY_SQLSTR == generated_fiscalhour_sqlstr
-
-        # WHEN
-        mont_tablename = x_objs.mont_stage_tablename
-        focus_columns = {fiscal_title_str(), month_title_str()}
-        generated_fiscalmont_sqlstr = create_inconsistency_query(
-            fiscal_db_conn, mont_tablename, focus_columns, exclude_cols
-        )
-        # THEN
-        print(f"{generated_fiscalmont_sqlstr=}")
-        assert FISCALMONT_INCONSISTENCY_SQLSTR == generated_fiscalmont_sqlstr
-
-        # WHEN
-        week_tablename = x_objs.week_stage_tablename
-        focus_columns = {fiscal_title_str(), weekday_title_str()}
-        generated_fiscalweek_sqlstr = create_inconsistency_query(
-            fiscal_db_conn, week_tablename, focus_columns, exclude_cols
-        )
-        # THEN
-        print(f"{generated_fiscalweek_sqlstr=}")
-        assert FISCALWEEK_INCONSISTENCY_SQLSTR == generated_fiscalweek_sqlstr
+        for x_category, x_sqlstr in get_inconsistency_sqlstrs().items():
+            x_tablename = f"{x_category}_staging"
+            cat_config = idea_config.get(x_category)
+            cat_focus_columns = set(cat_config.get("jkeys").keys())
+            cat_focus_columns.remove(event_int_str())
+            cat_focus_columns.remove(face_name_str())
+            generated_cat_sqlstr = create_inconsistency_query(
+                conn, x_tablename, cat_focus_columns, exclude_cols
+            )
+            assert x_sqlstr == generated_cat_sqlstr
+            print(f"{x_category} checked...")
 
 
 def test_set_fiscal_staging_error_message_Scenario0_fiscalunit_WithNo_error_message(
